fix: move to new folder structure
