+++ packages/shared/src/lib/auxiliary/wallet-connect/constants/index.ts
