+++ packages/shared/src/lib/auxiliary/wallet-connect/constants/supported-events.constant.ts
