+++ packages/shared/src/lib/auxiliary/wallet-connect/constants/supported-methods.constant.ts
