+++ packages/shared/src/lib/auxiliary/wallet-connect/constants/wallet-metadata.constant.ts
