+++ packages/shared/src/lib/auxiliary/wallet-connect/handlers/eth_sign.handler.ts
