+++ packages/shared/src/lib/auxiliary/wallet-connect/handlers/eth_signTransaction.handler.ts
