+++ packages/shared/src/lib/auxiliary/wallet-connect/handlers/eth_signTypedData.handler.ts
