+++ packages/shared/src/lib/auxiliary/wallet-connect/handlers/index.ts
