+++ packages/shared/src/lib/auxiliary/wallet-connect/handlers/onSessionProposal.handler.ts
@@ -1,7 +1,7 @@
 import { Web3WalletTypes } from '@walletconnect/web3wallet'
 import { SUPPORTED_EVENTS, SUPPORTED_METHODS } from '../constants'
-import { getAllEvmAddresses } from '../utils'
 import { buildApprovedNamespaces } from '@walletconnect/utils'
+import { getAllEvmAddresses } from '../utils'
 import { getWalletClient } from '../stores'
 
 export function onSessionProposal(sessionProposal: Web3WalletTypes.SessionProposal): void {
