+++ packages/shared/src/lib/auxiliary/wallet-connect/handlers/onSessionRequest.handler.ts
