+++ packages/shared/src/lib/auxiliary/wallet-connect/index.ts
