+++ packages/shared/src/lib/auxiliary/wallet-connect/stores/index.ts
