+++ packages/shared/src/lib/auxiliary/wallet-connect/stores/wallet-client.store.ts
