+++ packages/shared/src/lib/auxiliary/wallet-connect/utils/getAllEvmAddresses.ts
