+++ packages/shared/src/lib/auxiliary/wallet-connect/utils/index.ts
