+++ packages/shared/src/lib/auxiliary/wallet-connect/utils/initializeWalletConnect.ts
