+++ packages/shared/src/lib/auxiliary/wallet-connect/utils/pairWithNewDApp.ts
