+++ packages/shared/src/lib/core/profile/actions/active-profile/login.ts
@@ -48,7 +48,7 @@ import { logout } from './logout'
 import { subscribeToWalletApiEventsForActiveProfile } from './subscribeToWalletApiEventsForActiveProfile'
 import { checkAndUpdateActiveProfileNetwork } from './checkAndUpdateActiveProfileNetwork'
 import { checkAndRemoveProfilePicture } from './checkAndRemoveProfilePicture'
-import { initializeWalletConnect } from '@lib/walletconnect'
+import { initializeWalletConnect } from '@auxiliary/wallet-connect'
 
 export async function login(loginOptions?: ILoginOptions): Promise<void> {
     const loginRouter = get(routerManager).getRouterForAppContext(AppContext.Login)
